requirements.txt
# Google ADK and AI dependencies
google-generativeai>=0.3.0

# Web framework
streamlit>=1.29.0
streamlit-option-menu>=0.3.2

# Charts and visualization
plotly>=5.17.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
loguru>=0.7.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0